:truck: Mv common funcs into misc module
